Moved image loading to separat class
